+++ CHANGELOG.md
@@ -13,6 +13,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Minimal recommended `Scarb` version is now `2.12.2` (updated from `2.11.4`)
 
+### Cast
+
+#### Removed
+
+- The deprecated `argent` option for `--type` flag in `account create` and `account import` commands. Use `ready` instead
+
 ## [0.54.0] - 2025-12-09
 
 ### Forge